Release version 2.0.1
Administriva: Update builds to use Gradle 4.2.0

Administriva: Switch builds to stop pulling dependencies from
              JCenter, as it's EOL

Cleanup: Convert the old mkresources.py script to Python 3

Bugfix: Update content definitions for Slylandro, Probes, and Safe
        Ones to match the 0.8.x content packs

Bugfix: Wallpaper offset is now remembered across device restarts

Trivial Java cleanups (null pointer checks, remove useless casts, etc.)

+++ UQMLiveWallpaper/src/main/java/net/submedia/android/uqmwallpaper/Content.java
@@ -210,15 +210,15 @@ public class Content {
             }
 
             Hotspot(String x, String y) {
-                this((int) Integer.parseInt(x), (int) Integer.parseInt(y));
+                this(Integer.parseInt(x), Integer.parseInt(y));
             }
 
             Hotspot(int x, String y) {
-                this(x, (int) Integer.parseInt(y));
+                this(x, Integer.parseInt(y));
             }
 
             Hotspot(String x, int y) {
-                this((int) Integer.parseInt(x), y);
+                this(Integer.parseInt(x), y);
             }
 
             @Override
